- Added PongGameView-class and implemented existing code.
