+++ app/src/main/java/com/example/racoonsquash/PongGameView.kt
@@ -1,0 +1,85 @@
+package com.example.racoonsquash
+
+import android.content.Context
+import android.graphics.Canvas
+import android.graphics.Color
+import android.graphics.Paint
+import android.graphics.Path
+import android.view.SurfaceHolder
+import android.view.SurfaceView
+
+class PongGameView (context: Context) : SurfaceView(context), SurfaceHolder.Callback, Runnable{
+
+    private var lineColor: Paint
+    private var leftBoundaryPath: Path? = null
+    private var rightBoundaryPath: Path? = null
+
+    var mHolder: SurfaceHolder? = holder
+
+    init {
+        if (mHolder != null) {
+            mHolder?.addCallback(this)
+        }
+
+        lineColor = Paint().apply {
+            color = Color.CYAN
+            style = Paint.Style.STROKE
+            strokeWidth = 10f
+        }
+    }
+
+    override fun surfaceCreated(holder: SurfaceHolder) {
+
+    }
+
+    override fun surfaceChanged(holder: SurfaceHolder, format: Int, width: Int, height: Int) {
+        leftBoundaryPath = createBoundaryPathLeft(width, height)
+        rightBoundaryPath = createBoundaryPathRight(width, height)
+        drawGameBounds(holder)
+
+    }
+
+    override fun surfaceDestroyed(holder: SurfaceHolder) {
+
+    }
+
+    override fun run() {
+
+    }
+
+    fun drawGameBounds(holder: SurfaceHolder) {
+        val canvas: Canvas? = holder.lockCanvas()
+
+        canvas?.drawColor(Color.BLACK)
+
+        rightBoundaryPath?.let {
+            canvas?.drawPath(it, lineColor)
+        }
+
+        leftBoundaryPath?.let {
+            canvas?.drawPath(it, lineColor)
+        }
+
+        holder.unlockCanvasAndPost(canvas)
+    }
+
+    // Enbart spelplan med linje för syns skull, vänster sidolinje.
+    private fun createBoundaryPathLeft(width: Int, height: Int): Path {
+        val pathLeft = Path()
+
+        pathLeft.moveTo(0f, 0f)
+        pathLeft.lineTo(0f, height.toFloat())
+
+        return pathLeft
+    }
+
+    // Enbart spelplan med linje för syns skull, höger sidolinje.
+    private fun createBoundaryPathRight(width: Int, height: Int): Path {
+        val pathRight = Path()
+
+        pathRight.moveTo(width.toFloat(), 0f)
+        pathRight.lineTo(width.toFloat(), height.toFloat())
+
+        return pathRight
+    }
+}
